# Perf object detection