## Summary
This PR partially addresses #862 

> 🎯 **Purpose**: Improve performance of find_label_issues in object_detection/filter.py file.
>

**[ ✏️ Write your summary here. ]**
A significant improvement comes from vectorizing `_get_iou`, `_euc_dis` and `_calculate_average_precision`. I searched for all references to the deleted `_mod_coordinates` and `_get_overlap` but they are not used anywhere else in the code. I assumed it is ok to remove them because they are not part of the public api however let me know if you prefer to deprecate them or keep both versions.


For memory I used the memory-profiler library. The code I used for benchmarking is copied below. In addition I sorted the imports in the modified files.

**Code Setup**
```python
import random

import numpy as np

from cleanlab.object_detection.filter import find_label_issues

np.random.seed(0)
random.seed(0)

N = 5_000
SIZE = 256
MAX_L = 10
MAX_M = 100
K = 20

# Create input data
labels = []
predictions = []
for _ in range(N):
    x = random.randint(1, MAX_L)
    bboxes = np.empty((x, 4))
    bboxes[:, :2] = np.random.randint(0, SIZE // 2, size=(x, 2))
    bboxes[:, 2:] = np.random.randint(SIZE // 2, SIZE, size=(x, 2))
    img_labels = np.random.randint(K, size=x)
    label = {"bboxes": bboxes, "labels": img_labels}
    labels.append(label)

    y = random.randint(x, MAX_M)
    prediction = np.empty((K, y, 5))
    prediction[:, :, :2] = np.random.randint(0, SIZE // 2, size=(K, y, 2))
    prediction[:, :, 2:4] = np.random.randint(SIZE // 2, SIZE, size=(K, y, 2))
    prediction[:, :, 4] = np.random.random((K, y))
    predictions.append(prediction)
```

**Current version**
```python
%%timeit
%memit find_label_issues(labels, predictions)
# peak memory: 1317.51 MiB, increment: 820.18 MiB
# peak memory: 1317.58 MiB, increment: 0.07 MiB
# peak memory: 1317.60 MiB, increment: 0.02 MiB
# peak memory: 1317.86 MiB, increment: 0.28 MiB
# peak memory: 1319.04 MiB, increment: 1.18 MiB
# peak memory: 1319.09 MiB, increment: 0.07 MiB
# peak memory: 1319.09 MiB, increment: 0.02 MiB
# peak memory: 1319.11 MiB, increment: 0.02 MiB
# 8min 7s ± 30.2 s per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

**This PR**
```python
%%timeit
%memit find_label_issues(labels, predictions)
# peak memory: 1291.41 MiB, increment: 792.09 MiB
# peak memory: 1291.82 MiB, increment: 1.19 MiB
# peak memory: 1291.87 MiB, increment: 0.05 MiB
# peak memory: 1292.05 MiB, increment: 0.18 MiB
# peak memory: 1292.07 MiB, increment: 0.00 MiB
# peak memory: 1292.07 MiB, increment: 0.00 MiB
# peak memory: 1292.07 MiB, increment: 0.00 MiB
# peak memory: 1292.07 MiB, increment: 0.00 MiB
# 29.2 s ± 574 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```


## Testing

> 🔍 Testing Done: Existing test suite and output verification (explained below).


## References


## Reviewer Notes
> 💡 Include any specific points for the reviewer to consider during their review.

I was not sure about which input would be more realistic, I have tried different options but the improvements were similar, however let me know if you want me to test a specific combination of inputs, I tried to use the largest input dataset I could while still being able to run the master %%timeit in about an hour.

I have noticed that the `num_procs` argument in `_calculate_ap_per_class` is always set to the default value. Thus, i made a few changes there to ensure that the pool is only created when `num_procs` is higher than 1.

After each change I ran the tests and they were catching any minor mistake I made. However once they were passing I ensured with some random input that both versions (master and this PR) still produced the same exact output.